web/roster-member Extract the Region model in a separate file.
