+++ ui/web/src/app/wow/character-card/character-card.component.scss
@@ -22,7 +22,6 @@
 
 $card-container-height: 36px;
 $card-shadow-backdrop: 2px;
-$card-additional-info-container-height: 36px;
 
 .character-card-container {
   display: flex;
